[wip] to fix networking and add experiments (#1)
* it worked!!

The issue (at least from this test) was creating the
instances in different subsets. That is OK if we can use
DNS from the router, but not if we are using the internal
addresses (as userentes is). The fix was still to create
3 subnets / AZones (it will get angry if you do not) but
to only specify one for the autoscaling group! Could it
have been luck? Maybe - but 7 working seems like it is not

* add mnist example and refactor for arm

using the amd64 options that do not have efa and single
threading really does not work, so everything is refactored
for the hpc instances with efa. phew.

Signed-off-by: vsoch <vsoch@users.noreply.github.com>