efa is working with flux and usernetes
This is a tiny win, and albeit it is slow, lammps
is fully working with efa in the flux operator usernetes.
My cluster was too small for 2 nodes (and the control
plane was tainted) so I will need to account for that and
bring up another node, and will do this next time. It is
really OK that the usernetes setup is slow, we expected
that, and it provides a baseline for improvement. We do
need to create another experiment that again demonstrates
the integration of the two environments, and I will need
to think about that more.

Signed-off-by: vsoch <vsoch@users.noreply.github.com>